refactor: rewrite layout as flex
Closes #86
