+++ src/client/components/Footer/index.ts
