+++ src/client/components/Layout/Layout.jsx
@@ -1,19 +1,23 @@
-import { useStore } from "@nanostores/react";
-
+//import { useStore } from "@nanostores/react";
 import styles from "./Layout.module.css";
-import { uiStore } from "@stores/uiStore";
+//import { uiStore } from "@stores/uiStore";
+import Header from "@components/Header";
+import Loading from "@components/Loading";
 import Navigation from "@components/Navigation";
+import Footer from "@components/Footer";
+import SettingsMenu from "@components/SettingsMenu";
 
-export default function Layout({ children }) {
-  const { mobileOpen } = useStore(uiStore);
-
+export default function Layout({ children, loading = false }) {
   return (
-    <div className={styles.wrapper}>
-      {mobileOpen && <Navigation />}
+    <>
+      <Header />
       <div className={styles.inner}>
         <Navigation />
-        <main>{children}</main>
+        <main>{loading ? null : children}</main>
       </div>
-    </div>
+      <Footer />
+      <SettingsMenu />
+      {loading && <Loading /> /* TODO: replace with Pending in .inner */}
+    </>
   );
 }
